+++ CHANGES.md
@@ -21,6 +21,11 @@
   requests.
 * S3Source supports multiple endpoints when using ScriptLookupStrategy.
 
+### Processors
+* Updated Apache PDFBox to version 3.0.3.
+* PdfBoxProcessor property added to reduce memory during render
+    * `processor.pdf.sub_sampling_enabled` Boolean. Defaults to FALSE. Allows the PDFBox Page renderer to discard lines, saving memory by reducing quality.
+
 ### Caches
 
 * S3Cache uses multipart uploads, which reduces memory usage when caching